Testing the window functions

+++ src/main/java/ee/ut/cs/dsg/StreamCardinality/ApproximateAggregateFunction/AdaptiveCountingAggregationFunction.java
@@ -4,7 +4,7 @@ package ee.ut.cs.dsg.StreamCardinality.ApproximateAggregateFunction;
 import org.apache.flink.api.common.functions.AggregateFunction;
 import org.apache.flink.api.java.tuple.Tuple3;
 
-public class AdaptiveCountingAggregationFunction implements AggregateFunction<Tuple3<Long, String, Integer>, AdaptiveCountingAccumulator, Tuple3<Long,String,Integer>>{
+public class AdaptiveCountingAggregationFunction implements AggregateFunction<Tuple3<Long, String, Long>, AdaptiveCountingAccumulator, Tuple3<Long,String,Long>>{
     public AdaptiveCountingAccumulator createAccumulator() { return new AdaptiveCountingAccumulator(); }
 
     public AdaptiveCountingAccumulator merge(AdaptiveCountingAccumulator a, AdaptiveCountingAccumulator b) {
@@ -12,7 +12,7 @@ public class AdaptiveCountingAggregationFunction implements AggregateFunction<Tu
         return a;
     }
 
-    public AdaptiveCountingAccumulator add(Tuple3<Long, String, Integer> value, AdaptiveCountingAccumulator acc) {
+    public AdaptiveCountingAccumulator add(Tuple3<Long, String, Long> value, AdaptiveCountingAccumulator acc) {
         acc.f0 = value.f0;
         acc.f1 = value.f1;
         long val = Math.round(value.f2);
@@ -21,12 +21,12 @@ public class AdaptiveCountingAggregationFunction implements AggregateFunction<Tu
         return acc;
     }
 
-    public Tuple3<Long, String, Integer> getResult(AdaptiveCountingAccumulator acc) {
-        Tuple3<Long,String,Integer> res= new Tuple3<>();
+    public Tuple3<Long, String, Long> getResult(AdaptiveCountingAccumulator acc) {
+        Tuple3<Long,String,Long> res= new Tuple3<>();
         res.f0 = acc.f0;
         res.f1 = acc.f1;
         try{
-            res.f2 = (int) acc.acc.cardinality();
+            res.f2 = acc.acc.cardinality();
         } catch (Exception e) {
             e.printStackTrace();
         }
